+++ pkgs/ffigen/lib/src/executables/ffigen.dart
@@ -6,13 +6,27 @@
 import 'dart:io';
 
 import 'package:args/args.dart';
+import 'package:cli_util/cli_logging.dart' show Ansi;
 import 'package:logging/logging.dart';
 import 'package:package_config/package_config.dart';
 import 'package:yaml/yaml.dart' as yaml;
 
 import '../../ffigen.dart';
 
-final _logger = Logger('ffigen.ffigen');
+final _ansi = Ansi(Ansi.terminalSupportsAnsi);
+final logger = () {
+  final logger = Logger('ffigen.ffigen');
+  logger.onRecord.listen((record) {
+    final levelStr = '[${record.level.name}]'.padRight(9);
+    final log = '$levelStr: ${record.message}';
+    if (record.level < Level.SEVERE) {
+      print(log);
+    } else {
+      print('${_ansi.red}$log${_ansi.none}');
+    }
+  });
+  return logger;
+}();
 
 const compilerOpts = 'compiler-opts';
 const ignoreSourceErrors = 'ignore-source-errors';
@@ -32,34 +46,35 @@ Future<void> main(List<String> args) async {
   // Parses the cmd args. This will print usage and exit if --help was passed.
   final argResult = getArgResults(args);
 
-  final ffigen = FfiGen(logLevel: _parseLogLevel(argResult));
+  Logger.root.level = _parseLogLevel(argResult);
 
   // Create a config object.
-  Config config;
+  FfiGen config;
   try {
     config = getConfig(argResult, await findPackageConfig(Directory.current));
   } on FormatException {
-    _logger.severe('Please fix configuration errors and re-run the tool.');
+    logger.severe('Please fix configuration errors and re-run the tool.');
     exit(1);
   }
 
-  ffigen.run(config);
+  config.generate(logger);
 }
 
-Config getConfig(ArgResults result, PackageConfig? packageConfig) {
-  _logger.info('Running in ${Directory.current}');
+FfiGen getConfig(ArgResults result, PackageConfig? packageConfig) {
+  logger.info('Running in ${Directory.current}');
   YamlConfig config;
 
   // Parse config from yaml.
-  if (result.wasParsed(conf)) {
-    config = getConfigFromCustomYaml(result[conf] as String, packageConfig);
+  final configOption = (result[conf] as String?) ?? result.rest.firstOrNull;
+  if (configOption != null) {
+    config = getConfigFromCustomYaml(configOption, packageConfig);
   } else {
     config = getConfigFromPubspec(packageConfig);
   }
 
   // Add compiler options from command line.
   if (result.wasParsed(compilerOpts)) {
-    _logger.fine('Passed compiler opts - "${result[compilerOpts]}"');
+    logger.fine('Passed compiler opts - "${result[compilerOpts]}"');
     config.addCompilerOpts(result[compilerOpts] as String, highPriority: true);
   }
 
@@ -77,7 +92,7 @@ YamlConfig getConfigFromPubspec(PackageConfig? packageConfig) {
   final pubspecFile = File(pubspecName);
 
   if (!pubspecFile.existsSync()) {
-    _logger.severe(
+    logger.severe(
       'Error: $pubspecName not found, please run this tool from '
       'the root of your package.',
     );
@@ -92,7 +107,7 @@ YamlConfig getConfigFromPubspec(PackageConfig? packageConfig) {
   final bindingsConfigMap = pubspecYaml[configKey] as yaml.YamlMap?;
 
   if (bindingsConfigMap == null) {
-    _logger.severe("Couldn't find an entry for '$configKey' in $pubspecName.");
+    logger.severe("Couldn't find an entry for '$configKey' in $pubspecName.");
     exit(1);
   }
   return YamlConfig.fromYaml(
@@ -110,7 +125,7 @@ YamlConfig getConfigFromCustomYaml(
   final yamlFile = File(yamlPath);
 
   if (!yamlFile.existsSync()) {
-    _logger.severe('Error: $yamlPath not found.');
+    logger.severe('Error: $yamlPath not found.');
     exit(1);
   }