+++ pkgs/ffigen/lib/src/header_parser/sub_parsers/functiondecl_parser.dart
@@ -7,14 +7,14 @@ import 'package:logging/logging.dart';
 import '../../code_generator.dart';
 import '../../config_provider/config_types.dart';
 import '../clang_bindings/clang_bindings.dart' as clang_types;
-import '../data.dart';
 import '../utils.dart';
 import 'api_availability.dart';
 
-final _logger = Logger('ffigen.header_parser.functiondecl_parser');
-
 /// Parses a function declaration.
-List<Func> parseFunctionDeclaration(clang_types.CXCursor cursor) {
+List<Func> parseFunctionDeclaration(
+  Context context,
+  clang_types.CXCursor cursor,
+) {
   /// Multiple values are since there may be more than one instance of the
   /// same base C function with different variadic arguments.
   final funcs = <Func>[];
@@ -24,7 +24,7 @@ List<Func> parseFunctionDeclaration(clang_types.CXCursor cursor) {
 
   final apiAvailability = ApiAvailability.fromCursor(cursor);
   if (apiAvailability.availability == Availability.none) {
-    _logger.info('Omitting deprecated function $funcName');
+    logger.info('Omitting deprecated function $funcName');
     return funcs;
   }
 
@@ -33,7 +33,7 @@ List<Func> parseFunctionDeclaration(clang_types.CXCursor cursor) {
   if (cachedFunc != null) {
     funcs.add(cachedFunc);
   } else {
-    _logger.fine('++++ Adding Function: ${cursor.completeStringRepr()}');
+    logger.fine('++++ Adding Function: ${cursor.completeStringRepr()}');
 
     final returnType = cursor.returnType().toCodeGenType();
 
@@ -44,13 +44,13 @@ List<Func> parseFunctionDeclaration(clang_types.CXCursor cursor) {
     for (var i = 0; i < totalArgs; i++) {
       final paramCursor = clang.clang_Cursor_getArgument(cursor, i);
 
-      _logger.finer('===== parameter: ${paramCursor.completeStringRepr()}');
+      logger.finer('===== parameter: ${paramCursor.completeStringRepr()}');
 
       final paramType = paramCursor.toCodeGenType();
       if (paramType.isIncompleteCompound) {
         incompleteStructParameter = true;
       } else if (paramType.baseType is UnimplementedType) {
-        _logger.finer('Unimplemented type: ${paramType.baseType}');
+        logger.finer('Unimplemented type: ${paramType.baseType}');
         unimplementedParameterType = true;
       }
 
@@ -73,11 +73,11 @@ List<Func> parseFunctionDeclaration(clang_types.CXCursor cursor) {
     if (clang.clang_Cursor_isFunctionInlined(cursor) != 0 &&
         clang.clang_Cursor_getStorageClass(cursor) !=
             clang_types.CX_StorageClass.CX_SC_Extern) {
-      _logger.fine(
+      logger.fine(
         '---- Removed Function, reason: inline function: '
         '${cursor.completeStringRepr()}',
       );
-      _logger.warning(
+      logger.warning(
         "Skipped Function '$funcName', inline functions are not supported.",
       );
       // Returning empty so that [addToBindings] function excludes this.
@@ -85,11 +85,11 @@ List<Func> parseFunctionDeclaration(clang_types.CXCursor cursor) {
     }
 
     if (returnType.isIncompleteCompound || incompleteStructParameter) {
-      _logger.fine(
+      logger.fine(
         '---- Removed Function, reason: Incomplete struct pass/return by '
         'value: ${cursor.completeStringRepr()}',
       );
-      _logger.warning(
+      logger.warning(
         "Skipped Function '$funcName', Incomplete struct pass/return by "
         'value not supported.',
       );
@@ -99,11 +99,11 @@ List<Func> parseFunctionDeclaration(clang_types.CXCursor cursor) {
 
     if (returnType.baseType is UnimplementedType ||
         unimplementedParameterType) {
-      _logger.fine(
+      logger.fine(
         '---- Removed Function, reason: unsupported return type or '
         'parameter type: ${cursor.completeStringRepr()}',
       );
-      _logger.warning(
+      logger.warning(
         "Skipped Function '$funcName', function has unsupported return type "
         'or parameter type.',
       );
@@ -122,7 +122,7 @@ List<Func> parseFunctionDeclaration(clang_types.CXCursor cursor) {
       if (clang.clang_isFunctionTypeVariadic(cursor.type()) == 1) {
         varArgFunctions = config.varArgFunctions[funcName]!;
       } else {
-        _logger.warning(
+        logger.warning(
           'Skipping variadic-argument config for function '
           "'$funcName' since its not variadic.",
         );