+++ pkgs/ffigen/lib/src/header_parser/utils.dart
@@ -9,12 +9,10 @@ import 'package:logging/logging.dart';
 import 'package:pub_semver/pub_semver.dart';
 
 import '../code_generator.dart';
+import '../context.dart';
 import '../config_provider/config_types.dart';
 import 'clang_bindings/clang_bindings.dart' as clang_types;
-import 'data.dart';
 import 'type_extractor/extractor.dart';
-
-final _logger = Logger('ffigen.header_parser.utils');
 
 const exceptionalVisitorReturn =
     clang_types.CXChildVisitResult.CXChildVisit_Break;
@@ -25,7 +23,7 @@ typedef _CursorVisitorCallback =
 /// Logs the warnings/errors returned by clang for a translation unit.
 void logTuDiagnostics(
   Pointer<clang_types.CXTranslationUnitImpl> tu,
-  Logger logger,
+  Context context,
   String header, {
   Level logLevel = Level.SEVERE,
 }) {
@@ -33,7 +31,10 @@ void logTuDiagnostics(
   if (total == 0) {
     return;
   }
-  logger.log(logLevel, 'Header $header: Total errors/warnings: $total.');
+  context.logger.log(
+    logLevel,
+    'Header $header: Total errors/warnings: $total.',
+  );
   for (var i = 0; i < total; i++) {
     final diag = clang.clang_getDiagnostic(tu, i);
     if (clang.clang_getDiagnosticSeverity(diag) >=
@@ -50,12 +51,35 @@ void logTuDiagnostics(
               .CXDiagnosticDisplayOptions
               .CXDiagnostic_DisplayCategoryName,
     );
-    logger.log(logLevel, '    ${cxstring.toStringAndDispose()}');
+    context.logger.log(logLevel, '    ${cxstring.toStringAndDispose()}');
     clang.clang_disposeDiagnostic(diag);
   }
 }
 
-extension CXSourceRangeExt on Pointer<clang_types.CXSourceRange> {
+extension CXSourceLocationExt on clang_types.CXSourceLocation {
+  (String, int) get fileAndOffset {
+    final filePtr = calloc<Pointer<Void>>();
+    final offsetPtr = calloc<UnsignedInt>();
+
+    clang.clang_getFileLocation(this, filePtr, nullptr, nullptr, offsetPtr);
+    final file = clang.clang_getFileName(filePtr.value).toStringAndDispose();
+    final offset = offsetPtr.value;
+
+    calloc.free(filePtr);
+    calloc.free(offsetPtr);
+
+    return (file, offset);
+  }
+}
+
+extension CXSourceRangeExt on clang_types.CXSourceRange {
+  clang_types.CXSourceLocation get start => clang.clang_getRangeStart(this);
+  clang_types.CXSourceLocation get end => clang.clang_getRangeEnd(this);
+  ((String, int), (String, int)) toTuple() =>
+      (start.fileAndOffset, end.fileAndOffset);
+}
+
+extension CXSourceRangePtrExt on Pointer<clang_types.CXSourceRange> {
   void dispose() {
     calloc.free(this);
   }
@@ -250,15 +274,13 @@ extension CXCursorExt on clang_types.CXCursor {
 const commentPrefix = '/// ';
 const nesting = '  ';
 
-/// Stores the [clang_types.CXSourceRange] of the last comment.
-clang_types.CXSourceRange? lastCommentRange;
-
 /// Returns a cursor's associated comment.
 ///
 /// The given string is wrapped at line width = 80 - [indent]. The [indent] is
 /// [commentPrefix].length by default because a comment starts with
 /// [commentPrefix].
 String? getCursorDocComment(
+  Context context,
   clang_types.CXCursor cursor, {
   int indent = commentPrefix.length,
   String? fallbackComment,
@@ -267,10 +289,8 @@ String? getCursorDocComment(
   String? formattedDocComment;
   final currentCommentRange = clang.clang_Cursor_getCommentRange(cursor);
 
-  // See if this comment and the last comment both point to the same source
-  // range.
-  if (lastCommentRange != null &&
-      clang.clang_equalRanges(lastCommentRange!, currentCommentRange) != 0) {
+  // Only report the comment if we haven't reported this comment before.
+  if (!context.reportedCommentRanges.add(currentCommentRange.toTuple())) {
     formattedDocComment = null;
   } else {
     switch (config.commentType.length) {
@@ -289,7 +309,6 @@ String? getCursorDocComment(
         formattedDocComment = null;
     }
   }
-  lastCommentRange = currentCommentRange;
   final docs = [formattedDocComment ?? fallbackComment, availability].nonNulls;
   return docs.isEmpty ? null : docs.join('\n\n');
 }
@@ -533,7 +552,10 @@ class BindingsIndex {
 }
 
 class CursorIndex {
+  final Logger _logger;
   final _usrCursorDefinition = <String, clang_types.CXCursor>{};
+
+  CursorIndex(this._logger);
 
   /// Returns the Cursor definition (if found) or itself.
   clang_types.CXCursor getDefinition(clang_types.CXCursor cursor) {